+++ empire/server/core/db/models.py
@@ -582,6 +582,7 @@ class Bypass(Base):
     authors = Column(JSON)
     code = Column(Text)
     language = Column(String(255))
+    is_default = Column(Boolean, default=False, nullable=False)
     created_at = Column(UtcDateTime, nullable=False, default=utcnow())
     updated_at = Column(
         UtcDateTime, default=utcnow(), onupdate=utcnow(), nullable=False