+++ empire/server/stagers/windows/bunny.py
@@ -51,7 +51,7 @@ class Stager:
             "Bypasses": {
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
-                "Value": "mattifestation etw",
+                "Value": "",
             },
             "Language": {
                 "Description": "Language of the stager to generate.",