+++ empire/server/stagers/windows/wmic.py
@@ -83,7 +83,7 @@ class Stager:
             "Bypasses": {
                 "Description": "Bypasses as a space separated list to be prepended to the launcher",
                 "Required": False,
-                "Value": "mattifestation etw",
+                "Value": "",
             },
         }